# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/initiatives": {
      "post": {
        "operationId": "InitiativesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInitiativeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      }
    },
    "/api/v1/initiatives/{id}": {
      "patch": {
        "operationId": "InitiativesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInitiativeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiatives"
        ]
      }
    },
    "/api/v1/initiative-years": {
      "get": {
        "operationId": "InitiativeYearsController_list",
        "parameters": [
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiativeYearsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiative-years"
        ]
      }
    },
    "/api/v1/initiative-years/{id}": {
      "get": {
        "operationId": "InitiativeYearsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiativeYearResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiative-years"
        ]
      },
      "patch": {
        "operationId": "InitiativeYearsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInitiativeYearDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiative-years"
        ]
      },
      "delete": {
        "operationId": "InitiativeYearsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "revision",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiative-years"
        ]
      }
    },
    "/api/v1/initiative-years/{id}/backlog": {
      "patch": {
        "operationId": "InitiativeYearsController_updateBacklog",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBacklogDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiative-years"
        ]
      }
    },
    "/api/v1/initiative-years/{id}/preparation": {
      "patch": {
        "operationId": "InitiativeYearsController_updatePreparation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreparationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiative-years"
        ]
      }
    },
    "/api/v1/initiative-years/{id}/cards": {
      "post": {
        "operationId": "InitiativeYearsController_createCard",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuarterCardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiative-years"
        ]
      }
    },
    "/api/v1/initiative-years/extend": {
      "post": {
        "operationId": "InitiativeYearsController_extend",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendYearsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "initiative-years"
        ]
      }
    },
    "/api/v1/quarter-cards": {
      "get": {
        "operationId": "QuarterCardsController_list",
        "parameters": [
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quarter",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "view",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuarterCardsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quarter-cards"
        ]
      }
    },
    "/api/v1/quarter-cards/{id}": {
      "get": {
        "operationId": "QuarterCardsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuarterCardResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quarter-cards"
        ]
      },
      "patch": {
        "operationId": "QuarterCardsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quarter-cards"
        ]
      },
      "delete": {
        "operationId": "QuarterCardsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "revision",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quarter-cards"
        ]
      }
    },
    "/api/v1/quarter-cards/{id}/move": {
      "post": {
        "operationId": "QuarterCardsController_move",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeriodCommandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quarter-cards"
        ]
      }
    },
    "/api/v1/quarter-cards/{id}/continue": {
      "post": {
        "operationId": "QuarterCardsController_continueCard",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeriodCommandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quarter-cards"
        ]
      }
    },
    "/api/v1/quarter-cards/{cardId}/scope/{itemId}/move": {
      "post": {
        "operationId": "QuarterCardsController_moveScope",
        "parameters": [
          {
            "name": "cardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeriodCommandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quarter-cards"
        ]
      }
    },
    "/api/v1/quarter-cards/{cardId}/scope/{itemId}/copy": {
      "post": {
        "operationId": "QuarterCardsController_copyScope",
        "parameters": [
          {
            "name": "cardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeriodCommandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "quarter-cards"
        ]
      }
    },
...
  },
  "info": {
    "title": "PMO Hub API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiSuccessDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true,
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string",
            "example": "Операцію виконано"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "success"
        ]
      },
      "PreparationInputDto": {
        "type": "object",
        "properties": {
          "manager_id": {
            "type": "string",
            "format": "uuid"
          },
          "priority_id": {
            "type": "string",
            "format": "uuid"
          },
          "department_ids": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "department_ids"
        ]
      },
...
      "InitialQuarterCardDto": {
        "type": "object",
        "properties": {
          "manager_id": {
            "type": "string",
            "format": "uuid"
          },
          "priority_id": {
            "type": "string",
            "format": "uuid"
          },
          "department_ids": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "quarter": {
            "enum": [
              "Q1",
              "Q2",
              "Q3",
              "Q4"
            ],
            "type": "string"
          },
          "status_id": {
            "type": "string",
            "format": "uuid"
          },
          "notes": {
            "type": "string"
          },
          "custom_fields": {
            "type": "object",
            "additionalProperties": true
          },
          "scope": {
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateScopeItemDto"
            }
          }
        },
        "required": [
          "department_ids",
          "quarter",
          "scope"
        ]
      },
      "CreateInitiativeDto": {
        "type": "object",
        "properties": {
          "kind": {
            "enum": [
              "PROJECT",
              "OPERATIONAL_TASK"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "year": {
            "type": "number",
            "minimum": 2000,
            "maximum": 2200
          },
          "strategic_goal": {
            "type": "string"
          },
          "preparation": {
            "$ref": "#/components/schemas/PreparationInputDto"
          },
          "initial_card": {
            "$ref": "#/components/schemas/InitialQuarterCardDto"
          }
        },
        "required": [
          "kind",
          "name",
          "year",
          "preparation"
        ]
      },
      "UpdateInitiativeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "revision": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "name",
          "revision"
        ]
      },
...
      "InitiativeYearReadModelDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "initiative_id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "PROJECT",
              "OPERATIONAL_TASK"
            ]
          },
          "name": {
            "type": "string"
          },
          "initiative_revision": {
            "type": "number"
          },
          "year": {
            "type": "number"
          },
          "strategic_goal": {
            "type": "string",
            "nullable": true
          },
          "revision": {
            "type": "number"
          },
          "preparation": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PreparationStageReadModelDto"
              }
            ]
          },
          "cards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuarterCardSummaryDto"
            }
          }
        },
        "required": [
          "id",
          "initiative_id",
          "kind",
          "name",
          "initiative_revision",
          "year",
          "strategic_goal",
          "revision",
          "preparation",
          "cards"
        ]
      },
      "InitiativeYearsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InitiativeYearReadModelDto"
            }
          }
        },
        "required": [
          "success",
          "message",
          "data"
        ]
      },
      "InitiativeYearResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/InitiativeYearReadModelDto"
          }
        },
        "required": [
          "success",
          "message",
          "data"
        ]
      },
      "UpdateInitiativeYearDto": {
        "type": "object",
        "properties": {
          "strategic_goal": {
            "type": "string"
          },
          "revision": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "revision"
        ]
      },
      "UpdateBacklogDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "strategic_goal": {
            "type": "string"
          },
          "initiative_revision": {
            "type": "number",
            "minimum": 1
          },
          "year_revision": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "name",
          "initiative_revision",
          "year_revision"
        ]
      },
      "UpdatePreparationDto": {
        "type": "object",
        "properties": {
          "manager_id": {
            "type": "string",
            "format": "uuid"
          },
          "priority_id": {
            "type": "string",
            "format": "uuid"
          },
          "department_ids": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "revision": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "department_ids",
          "revision"
        ]
      },
      "CreateQuarterCardDto": {
        "type": "object",
        "properties": {
          "quarter": {
            "enum": [
              "Q1",
              "Q2",
              "Q3",
              "Q4"
            ],
            "type": "string"
          }
        },
        "required": [
          "quarter"
        ]
      },
      "RevisionTargetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "id",
          "revision"
        ]
      },
      "ExtendYearsDto": {
        "type": "object",
        "properties": {
          "source_years": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RevisionTargetDto"
            }
          },
          "target_year": {
            "type": "number",
            "minimum": 2000,
            "maximum": 2200
          }
        },
        "required": [
          "source_years",
          "target_year"
        ]
      },
...
      "QuarterCardsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuarterCardReadModelDto"
            }
          }
        },
        "required": [
          "success",
          "message",
          "data"
        ]
      },
      "QuarterCardResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/QuarterCardReadModelDto"
          }
        },
        "required": [
          "success",
          "message",
          "data"
        ]
      },
      "ScopeItemDto": {
        "type": "object",
        "properties": {
          "lineage_id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string"
          },
          "status_code": {
            "enum": [
              "DEFAULT",
              "GREEN",
              "YELLOW",
              "RED"
            ],
            "type": "string"
          },
          "weight_definition_id": {
            "type": "string",
            "format": "uuid"
          },
          "executor_department_ids": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "text",
          "status_code",
          "weight_definition_id",
          "executor_department_ids"
        ]
      },
      "UpdateCardDto": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "number",
            "minimum": 1
          },
          "manager_id": {
            "type": "string",
            "format": "uuid"
          },
          "priority_id": {
            "type": "string",
            "format": "uuid"
          },
          "department_ids": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "status_id": {
            "type": "string",
            "format": "uuid"
          },
          "notes": {
            "type": "string"
          },
          "custom_fields": {
            "type": "object",
            "additionalProperties": true
          },
          "scope": {
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScopeItemDto"
            }
          }
        },
        "required": [
          "revision",
          "department_ids",
          "status_id",
          "scope"
        ]
      },
      "PeriodCommandDto": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "number",
            "minimum": 1
          },
          "to_year": {
            "type": "number",
            "minimum": 2000,
            "maximum": 2200
          },
          "to_quarter": {
            "enum": [
              "Q1",
              "Q2",
              "Q3",
              "Q4"
            ],
            "type": "string"
          },
          "target_revision": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "revision",
          "to_year",
          "to_quarter"
        ]
      },
...
    }
  }
}